[Morello] Fix -morello-bounded-memargs=caller-only variadic edge case
If the caller is passing both fixed arguments in memory as well as
variadic ones then C9 does not point to the variadic arguments, as is
the case for -no-morello-bounded-memargs, but instead to the fixed
arguments in memory, with the variadic ones coming later. This means
that the callee needs to handle this edge case.

Note that this case makes -morello-bounded-memargs=caller-only slightly
incompatible with -no-morello-bounded-memargs. In practice it's rare and
deemed acceptable for CheriBSD to have this break between 22.12 and
23.11 (ABI breaks between releases are things users should expect, but
we do intend them to be narrow enough that you can upgrade a live system
without needing to consider them, so long as you don't skip a release),
with no good way to avoid it.

Fixes https://git.morello-project.org/morello/llvm-project/-/issues/77