[MC] Ditch .cfi_startproc purecap in favour of just .cfi_startproc
Countless assembly files out there use just .cfi_startproc, and it is
completely unnecessary, bug-prone and annoying to have to change each of
them to .cfi_startproc purecap just so Morello LLVM uses the right CFA
and return address information in its emitted DWARF. LLVM knows the ABI
for a whole host of other reasons, so .cfi_startproc should just do that
automatically, as it does already for CHERI-MIPS and CHERI-RISC-V.
Otherwise if you have a rogue .cfi_startproc without purecap (as is the
case today in CheriBSD, where none of our assembly files have it) you
end up with bogus CFI in the output, which helps nobody.

As part of this, the redundant C augmentation is removed, since it
should be implied by the ELF flags on the file in question, and GDB does
not look at it today. LLD retains support for it for now in order to
support linking old object files.

The existence of .cfi_startproc purecap likely dates back to before the
ABI was available to some of these functions, as I only added it
upstream when introducing CHERI-RISC-V (CHERI-MIPS has its own hacks).
Whilst there is theoretically a valid use case for having hybrid
functions in purecap binaries, neither the compiler nor the linker are
set up to support that today, so CFI support is the least of your
worries there. Moreover, purecap functions in a hybrid binary would also
be a thing you might want to support in just the same way. And, on top
of that, the prevailing design approach should be that making purecap
"just work" without any unnecessary changes, be it to assembly syntax or
DWARF output, is the priority, with weird mix-and-match use cases
fitting in around that with their own extensions.

Similarly, the existence of the augmentation likely dates back to before
there was an ELF flag to indicate the file is purecap (the LSB of the
entry point would have worked as a mostly-correct proxy for executables,
ignoring the case of a purecap kernel whose entry point called from
firmware is hybrid, but not for libraries), though one could have
inferred it still from the unwind info.

Note that .cfi_startproc purecap remains valid syntax for the GNU-style
asm parser, but will emit a warning and ignore the argument, just in
case Arm code is using it. The MASM parser does not retain this
compatibility as that is not of interest to Morello.