[AArch64][test] Extend TargetParserTest for Morello
A64C and C64 are mutually exclusive so need to be filtered out. One of
them perhaps should be checked, but the code also expects extensions and
features to be 1:1, which also isn't true ("+morello" is included on top
of either "+a64c"/"-c64" or "+c64"), so just leave the upstream test as
not testing Morello extensions.