[SelectionDAG] Fix for crash when building ICU4C
This fixes the crash found while building ICU4C where LLVM was using
NEON registers when lowering a memcpy from a zero constant and then
incorrectly triggering an assertion. No tests in this commit since
those will come as part of the next merge from CHERI LLVM (which had
the same underlying issue that just happened to not be reachable).