[Morello][lld][test] switch off now-broken MIPS test for llvm-15
With the clang-15 merge, CI pipeline now reports 1 fails in lld tests

lld/test/ELF/eh-frame-writable.test is a now-broken test.
This behaviour changed upstream, but we do not use MIPS so the
incorrect output does not matter.
For Morello LLVM, switch off this test.

In preparation for building with clang-15.
https://git.morello-project.org/morello/llvm-project/-/merge_requests/293

Change-Id: Id7f13b2dd0f15bade8f6d362534455a28e125025