[Morello][LLD] Never use function permissions for non-function symbols
This condition is a deviation from the CHERI-MIPS and CHERI-RISC-V
semantics: if a data symbol ends up being placed into an output section
with execute permission (e.g. constants placed in .text in a
hand-written assembly file, or a linker script merging .rodata into
.text, seen in OpenSSL's assembly routines and EDK2) then this changes
it from read-only to read-execute. The execute part is less of a problem
beyond giving unwanted permission, but being executable also means that
whatever processes the capability relocation will create a sentry,
breaking the ability to actually load from it, and thus rendering it
wholly broken for the code's intended use.

Fix this by dropping the misguided case and correctly handling GNU
IFUNCs, which are relying on this hack to pick up the right capability
permissions. This should only break incorrect code that fails to use
.type sym, %function, with the fix being to add such a line.