[Morello] Use TLI::getSimpleValueType for PointerTyID intrinsics
Asking for a capability like this via MVT::getType will become illegal
due to needing a valid DataLayout (and for any pointer type in general
upstream in future), so switch to asking TLI instead, which can be given
the DataLayout. These intrinsics could instead be made c128-only, like
how their integer counterparts are i64-only, but that's not how they've
been historically defined, so let's keep them the same. It would also be
nice to unify the two intrinsics anyway, at which point we'd still need
this due to using llvm_anyptr_ty.