[ELF][Morello] Stop supporting symbol size fallback from capinit data
If the symbol size for a capability relocation is not known, Morello
currently tries to read the size from the input section's capability
fragment. This is a hack, and the correct thing to do for such cases
would be to instead set the symbol size in the first place. Moreover,
Morello LLVM never emits such non-zero fragments, the only way to get
them is with hand-written assembly, and these days assembly should be
written to use .chericap rather than .capinit + data directives, which
does not even have syntax to express such a thing.

Therefore, stop honouring this field, and emit an error if an input is
seen that is trying to use it.