[ELF][Morello] Look at capinit fragment based on section not symbol
The intent of the original code is presumably to not try and look at the
capability fragment for a GOT entry, since there is no input data to
read from. The condition does catch that, but has the weird effect that
any symbol being present in the GOT prevents any other non-GOT
relocations referencing it from looking at the fragment (e.g. a global
initialised to point at the same value). We want to catch these too in
order to emit an error.

At the same time, there are other linker-generated capability
relocations than for data GOT entries; specifically, GotPltSection and
IgotPltSection are very similar cases, but do not in general add a data
GOT entry, so isInGot will report false and we'll attempt to dereference
a null data pointer. GotPltSection won't currently contain any
capability fragment relocations, since such symbols are non-preemptible
so don't require a PLT entry in the first place, but IgotPltSection will
in the case of a call to a non-preemptible IFUNC resolver.

To address both of these, switch the condition instead to ignoring all
relocations within synthetic sections.