[NFCI][Morello] Use proper capability fixup rather than hijacking FK_Data_8
A capinit is not an 8-byte fixup, it's a 16-byte one, and it's not for
plain data, it's for a capability. Follow other CHERI targets and define
a custom fixup rather than smuggling them in an FK_Data_8 annotated with
VK_CAPINIT.

This requires pushing emitCapInit's implementation down into the target
so it can use the target-specific fixup, as is the case with the normal
EmitCheriCapabilityImpl. This also stops using a no-op implementation in
MCStreamer itself as that should never be used.